% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{doc-common-rename}
\alias{doc-common-rename}
\title{Doc for rename}
\arguments{
\item{f}{A function or one-sided formula, interpreted by \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{...}{Passed into \code{f}.}
}
\value{
The renamed object.
}
\description{
Doc for rename
}
