% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel.R
\name{doc-common}
\alias{doc-common}
\title{Common documentation}
\arguments{
\item{x}{An object with names.}

\item{pattern}{A regular expression string (see \link{regex}).}
}
\description{
Common documentation
}
