% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_chunks.R
\name{name_chunks}
\alias{name_chunks}
\title{Name chunks in a single file}
\usage{
name_chunks(path)
}
\arguments{
\item{path}{}
}
\description{
Name unnamed chunks in a single file using the filename with extension stripped as basis.
}
\examples{
\dontrun{
file.copy(system.file("examples", "example1.Rmd", package = "namer"),
          "test.Rmd")
name_chunks("test.Rmd")
}
}
