% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_chunks.R
\name{name_dir_chunks}
\alias{name_dir_chunks}
\title{Name chunks of all Rmds in a dir}
\usage{
name_dir_chunks(dir)
}
\arguments{
\item{dir}{Path to folder}
}
\description{
Name unnamed chunks in a dir using the filenames with extension stripped as basis.
}
\details{
When using namer, please check the edits
 before pushing them to your code base. Such automatic
 chunk labelling is best paired with version control.
}
\examples{
temp_dir <- tempdir()
fs::dir_copy(system.file("examples", package = "namer"),
            temp_dir)
 # this is an example file that'd fail
fs::file_delete(file.path(temp_dir, "example4.Rmd"))
name_dir_chunks(temp_dir)
file.edit(file.path(temp_dir, "example1.Rmd"))
}
