% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named}
\alias{named}
\title{Subset objects by name}
\usage{
named(x, y)
}
\arguments{
\item{x}{An object with names.}

\item{y}{A vector of names.}
}
\value{
\code{setNames(x[y], y)}.
}
\description{
\code{x |> named(y)} is the same as \code{x[y]}, but
}
\details{
\itemize{
\item it makes your intent clearer;
\item it throws an error if \code{y} is not a character vector;
\item unmatched elements of \code{y} return \code{NA}, but with the name of
the element, whereas \code{x[y]} would return an \code{NA} name.
}
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
nms <- c("three", "two", "five")
vec |> named(nms)
# Compare to:
vec[nms]
}
