% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel.R
\name{named_in}
\alias{named_in}
\title{Select objects by name using a vector of matches}
\usage{
named_in(x, table)
}
\arguments{
\item{x}{A named vector.}

\item{table}{A vector.}
}
\value{
\preformatted{x[names(x) \%in\% y]
}
}
\description{
Select objects by name using a vector of matches
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
named_in(vec, c("one", "four"))

}
