% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named_starting}
\alias{named_starting}
\title{Subset objects by name using an initial substring}
\usage{
named_starting(x, prefix)
}
\arguments{
\item{x}{An object with names.}

\item{prefix}{A character string}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{x[startsWith(names(x), prefix)]
}\if{html}{\out{</div>}}
}
\description{
Subset objects by name using an initial substring
}
\examples{
vec <- c(one = 1, two = 2, three = 3, four = 4)
named_starting(vec, "t")
}
