% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_all}
\alias{rename_all}
\title{Rename all names}
\usage{
rename_all(x, f, ...)
}
\arguments{
\item{x}{An object with names.}

\item{f}{A function, one-sided formula, or character vector.}

\item{...}{Passed into \code{f}. An error is thrown if \code{...} is non-empty when \code{f}
is a character vector.}
}
\value{
The renamed object.
}
\description{
Rename all names
}
\details{
\itemize{
\item If \code{f} is a function it will be applied to the selected names. If it is
a formula and the 'rlang' package is installed, it will be converted to a
function by \code{\link[rlang:as_function]{rlang::as_function()}}, then
applied.
\item If \code{f} is a named character vector like \code{c(new_name = "old_name", ...)} then
\code{"old_name"} will become \code{"new_name"}, as in \code{dplyr::rename()}.
\item If \code{f} is an unnamed character vector, these will be the new names in order.
}
}
\examples{

vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_all(vec, tolower)

}
