% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel.R
\name{rename_chartr}
\alias{rename_chartr}
\alias{rename_tolower}
\alias{rename_toupper}
\title{Rename to upper or lower case or by a translation table}
\usage{
rename_chartr(x, old, new)

rename_tolower(x)

rename_toupper(x)
}
\arguments{
\item{x}{An object with names.}

\item{old, new}{Character strings passed to \code{\link[=chartr]{chartr()}}.}
}
\value{
\preformatted{setNames(x, tolower(names(x)))
}

or similar.
}
\description{
Rename to upper or lower case or by a translation table
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
rename_toupper(vec)
}
