% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel.R
\name{rename_fn}
\alias{rename_fn}
\title{Rename using a function}
\usage{
rename_fn(x, f, where = TRUE, ...)
}
\arguments{
\item{x}{An object with names.}

\item{f}{A function or one-sided formula, as in \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{where}{A set of indices for the names.}

\item{...}{Passed in to \code{f}.}
}
\value{
\code{x} after \code{names(x)[where] <- f(names(x)[where])}.
}
\description{
Rename using a function
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
rename_fn(vec, ~ substr(.x, 1, 2), where = 2:3)

}
