% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_like}
\alias{rename_like}
\title{Rename names that match a regular expression}
\usage{
rename_like(
  x,
  pattern,
  f,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{x}{An object with names.}

\item{pattern}{A regular expression string (see \link{regex}).}

\item{f}{A function, one-sided formula, or character vector.}

\item{...}{Passed into \code{f}. An error is thrown if \code{...} is non-empty when \code{f}
is a character vector.}

\item{ignore.case, perl, fixed, useBytes}{Passed into \code{\link[=grepl]{grepl()}}.}
}
\value{
The renamed object.
}
\description{
Rename names that match a regular expression
}
\details{
\itemize{
\item If \code{f} is a function it will be applied to the selected names. If it is
a formula and the 'rlang' package is installed, it will be converted to a
function by \code{\link[rlang:as_function]{rlang::as_function()}}, then
applied.
\item If \code{f} is a named character vector like \code{c(new_name = "old_name", ...)} then
\code{"old_name"} will become \code{"new_name"}, as in \code{dplyr::rename()}.
\item If \code{f} is an unnamed character vector, these will be the new names in order.
}
}
\examples{

vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_like(vec, "^T", paste0, "x")

}
