% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_lookup}
\alias{rename_lookup}
\title{Rename by looking up names in a table}
\usage{
rename_lookup(x, old, new, warn = FALSE)
}
\arguments{
\item{x}{An object with names.}

\item{old}{Character vector. Existing names will be found using
\code{match(names(x), old)}}

\item{new}{Character vector. A vector of new names to replace corresponding
elements in \code{old}.}

\item{warn}{Logical. Warn if any names are unmatched?}
}
\value{
\code{x} renamed according to \code{names(x) <- new[match(names(x), old)]}.
}
\description{
This is useful when you have a vector of old names and a vector of new names,
or columns in a data frame.
}
\details{
Unmatched names are left unchanged.
}
\examples{

df <- data.frame(
        old = c("One", "Two", "Three"),
        new = c("New", "Newer", "Newest")
      )
vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
vec |> rename_lookup(df$old, df$new)

}
