% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_match}
\alias{rename_match}
\title{Rename by matching existing names to a lookup table}
\usage{
rename_match(x, table, replacement, warn = FALSE)
}
\arguments{
\item{x}{An object with names.}

\item{table}{Character vector. Existing names will be found using
\code{match(names(x), table)}}

\item{replacement}{Character vector. A vector of names, indexed by matches
in \code{table}.}

\item{warn}{Logical. Warn if any names are unmatched?}
}
\value{
\code{x} renamed according to \code{names(x) <- replacement[match(names(x), table)]}.
}
\description{
This is useful when you have a column of old names and a column of new names.
}
\details{
Unmatched names are left unchanged.
}
\examples{

df <- data.frame(
        old = c("One", "Two", "Three"),
        new = c("New", "Newer", "Newest")
      )
vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_match(vec, df$old, df$new)

}
