% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_remove_prefix}
\alias{rename_remove_prefix}
\alias{rename_remove_suffix}
\title{Remove a prefix or suffix from names}
\usage{
rename_remove_prefix(x, prefix)

rename_remove_suffix(x, suffix)
}
\arguments{
\item{x}{An object with names.}

\item{prefix, suffix}{A length 1 character vector to remove.}
}
\value{
\code{x} with the prefix or suffix removed from \code{names(x)}.
}
\description{
Remove a prefix or suffix from names
}
\details{
\itemize{
\item If \code{f} is a function it will be applied to the selected names. If it is
a formula and the 'rlang' package is installed, it will be converted to a
function by \code{\link[rlang:as_function]{rlang::as_function()}}, then
applied.
\item If \code{f} is a named character vector like \code{c(new_name = "old_name", ...)} then
\code{"old_name"} will become \code{"new_name"}, as in \code{dplyr::rename()}.
\item If \code{f} is an unnamed character vector, these will be the new names in order.
}
}
\examples{

vec <- c("a.1" = 1, "aaa.1" = 2, "other" = 3, ".1" = 4)
rename_remove_suffix(vec, ".1")

vec <- c("x.a" = 1, "x.aaa" = 2, "other" = 3, "x." = 4)
rename_remove_prefix(vec, "x.")
}
