% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_starting}
\alias{rename_starting}
\title{Rename names that start with a prefix}
\usage{
rename_starting(x, prefix, f, ...)
}
\arguments{
\item{prefix}{A string.}

\item{f}{A function or one-sided formula, interpreted by \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{...}{Passed into \code{f}.}
}
\value{
The renamed object.
}
\description{
Rename names that start with a prefix
}
\examples{

vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_starting(vec, "T", gsub, "[aeiou]", "e")
}
