% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel.R
\name{rename_sub}
\alias{rename_sub}
\alias{rename_gsub}
\title{Rename using a regular expression}
\usage{
rename_sub(x, pattern, replacement, ...)

rename_gsub(x, pattern, replacement, ...)
}
\arguments{
\item{x}{An object with names.}

\item{pattern}{A regular expression string (see \link{regex}).}

\item{replacement}{A replacement string. See \code{\link[=sub]{sub()}}.}

\item{...}{Passed to \code{\link[=sub]{sub()}} or \code{\link[=gsub]{gsub()}}.}
}
\value{
\preformatted{setNames(x, sub(pattern, replacement, names(x), ...))
}
}
\description{
Rename using a regular expression
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
rename_sub(vec, "[aeiou]", "e")
rename_gsub(vec, "[aeiou]", "e")

}
