% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_subset}
\alias{rename_subset}
\title{Rename names indexed by a subset}
\usage{
rename_subset(x, matches, f, ...)
}
\arguments{
\item{matches}{A logical or numeric index.}

\item{f}{A function or one-sided formula, interpreted by \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{...}{Passed into \code{f}.}
}
\value{
The renamed object.
}
\description{
Rename names indexed by a subset
}
\examples{

vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_subset(vec, 2:3, paste0, 2:3)

}
