% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_by_name}
\alias{sort_by_name}
\title{Sort an object by its names}
\usage{
sort_by_name(x, decreasing = FALSE)
}
\arguments{
\item{x}{An object with names.}

\item{decreasing}{Logical. Should sort order be increasing or decreasing?}
}
\value{
\code{x[sort(names(x), decreasing = decreasing)]}
}
\description{
Sort an object by its names
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
sort_by_name(vec)
sort_by_name(vec, decreasing = TRUE)
}
