% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unname_all_chunks.R
\name{unname_all_chunks}
\alias{unname_all_chunks}
\title{Unname chunks in a single file}
\usage{
unname_all_chunks(path)
}
\arguments{
\item{path}{Path to file}
}
\description{
Unname all chunks except the setup chunk, in a single file
}
\examples{
\dontrun{
file.copy(system.file("examples", "example4.Rmd", package = "namer"),
          "test.Rmd")
unname_all_chunks("test.Rmd")
file.edit("test.Rmd")
}
}
