% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_chunks.R
\name{name_chunks}
\alias{name_chunks}
\title{Name chunks in a single file}
\usage{
name_chunks(path)
}
\arguments{
\item{path}{Path to file}
}
\description{
Name unnamed chunks in a single file using the filename with extension stripped as basis.
}
\details{
When using namer, please check the edits
 before pushing them to your code base. Such automatic
 chunk labelling is best paired with version control.
}
