% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named}
\alias{named}
\title{Subset objects by name}
\usage{
named(x, y)
}
\arguments{
\item{x}{An object with names.}

\item{y}{A character scalar.}
}
\value{
\preformatted{x[names(x) == y]
}
}
\description{
Subset objects by name
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
named(vec, "one")

}
