% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named_in}
\alias{named_in}
\title{Subset objects by name}
\usage{
named_in(x, y)
}
\arguments{
\item{x}{An object with names.}

\item{y}{A vector of names.}
}
\value{
\preformatted{x[intersect(names(x), y)]
}
}
\description{
Subset objects by name
}
\details{
\code{named_in(x, y)} is like \code{x[y]} except that:
\itemize{
\item unmatched elements of \code{y} do not return an element;
\item elements are returned in their original order within \code{x}.
}
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
named_in(vec, c("two", "one", "three", "five"))
}
