% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{named_in}
\alias{named_in}
\alias{not_named_in}
\title{Subset objects by name}
\usage{
named_in(x, y)

not_named_in(x, y)
}
\arguments{
\item{x}{An object with names.}

\item{y}{A vector of names.}
}
\value{
For \code{named_in}: \code{x[names(x) \%in\% y]}.

For \code{not_named_in}: \code{x[! names(x) \%in\% y]}.
}
\description{
Subset objects by name
}
\details{
\code{named_in(x, y)} is similar to \code{x[y]} except that:
\itemize{
\item unmatched elements of \code{y} do not return an \code{NA} element;
\item elements are returned in their original order within \code{x}.
}

\code{not_named_in(x, y)} returns elements of \code{x} whose name is not
an element of \code{y}.
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
vec |> named_in(c("two", "one", "three", "five"))
vec |> not_named_in(c("two", "three"))
}
