% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel.R
\name{named_like}
\alias{named_like}
\title{Select objects by name using a regular expression}
\usage{
named_like(x, pattern)
}
\arguments{
\item{x}{An object with names.}

\item{pattern}{A regular expression string (see \link{regex}).}
}
\value{
\preformatted{x[grepl(y, names(x))]
}
}
\description{
Select objects by name using a regular expression
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
named_like(vec, "^t")

}
