% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_in}
\alias{rename_in}
\title{Rename names \code{\%in\%} a set}
\usage{
rename_in(x, table, f, ...)
}
\arguments{
\item{table}{A character vector passed to \code{\%in\%}.}

\item{f}{A function or one-sided formula, interpreted by \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{...}{Passed into \code{f}.}
}
\value{
The renamed object.
}
\description{
Rename names \code{\%in\%} a set
}
\examples{

vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_in(vec, c("Two", "Three"), paste0, "x")

}
