% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_like}
\alias{rename_like}
\title{Rename names that match a regular expression}
\usage{
rename_like(
  x,
  pattern,
  f,
  ...,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{x}{An object with names.}

\item{pattern}{A regular expression string (see \link{regex}).}

\item{f}{A function or one-sided formula, interpreted by \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{...}{Passed into \code{f}.}

\item{ignore.case, perl, fixed, useBytes}{Passed into \code{\link[=grepl]{grepl()}}.}
}
\value{
The renamed object.
}
\description{
Rename names that match a regular expression
}
\examples{

vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_like(vec, "^T", paste0, "x")

}
