% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel.R
\name{rename_paste}
\alias{rename_paste}
\title{Rename using \code{paste()}}
\usage{
rename_paste(x, ..., sep = "")
}
\arguments{
\item{x}{An object with names.}

\item{...}{Character vectors to paste after the names.}

\item{sep}{Passed to \code{\link[=paste]{paste()}}. Note that the default is \code{""}.}
}
\value{
\preformatted{setNames(x, paste(names(x), ..., sep = sep))
}
}
\description{
Rename using \code{paste()}
}
\examples{

vec <- c(one = 1, two = 2, three = 3, four = 4)
rename_paste(vec, "X")

}
