% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_sub}
\alias{rename_sub}
\alias{rename_gsub}
\title{Rename using a regular expression}
\usage{
rename_sub(x, pattern, replacement, ...)

rename_gsub(x, pattern, replacement, ...)
}
\arguments{
\item{x}{An object with names.}

\item{pattern, replacement, ...}{Passed into \code{\link[=sub]{sub()}} or \code{\link[=gsub]{gsub()}}.}
}
\value{
The renamed object.
}
\description{
Rename using a regular expression
}
\details{
These functions always apply to all names.
}
\examples{

vec <- c("One" = 1, "Two" = 2, "Three" = 3, "Four" = 4)
rename_gsub(vec, "[aeiou]", "e")
rename_sub(vec, "([aeiou])", ".\\\\1.")
}
