% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel-package.R
\docType{package}
\name{rumpel-package}
\alias{rumpel}
\alias{rumpel-package}
\title{Manipulate objects by name}
\description{
Contains convenience functions for naming. Select subsets by 
  name using matches or regular expressions. Rename objects with regular
  expressions or paste.
}
\details{
\if{html}{\out{
<!-- README.md is generated from README.Rmd. Please edit that file -->
}}
\if{html}{\out{
<!-- badges: start -->
}}


\href{https://github.com/hughjonesd/rumpel/actions}{\figure{https://github.com/hughjonesd/rumpel/workflows/R-CMD-check/badge.svg}}\if{html}{\out{
<!-- badges: end -->
}}


In the fairy tale, Rumpelstiltskin could be controlled once the heroine
learned his name.

Rumpel is a tiny r package containing convenience functions for
manipulating objects by their names. Using these functions makes your
code easier to read, and reduces duplication:\if{html}{\out{<div class="r">}}\preformatted{library(rumpel)
library(magrittr)
my_object <- c(One = 1, Two = 2, Three = 3, Four = 4)

# Base R:
my_object[names(my_object) \%in\% c("One", "Two")]
#> One Two 
#>   1   2

# Nicer:
my_object \%>\% named_in(c("One", "Two"))
#> One Two 
#>   1   2


# Base R:
names(my_object)[names(my_object) \%in\% c("Two", "Three")] <-
      tolower(names(my_object)[names(my_object) \%in\% c("Two", "Three")])

# Nicer:
my_object \%>\% rename_in(c("Two", "Three"), tolower)
#>   One   two three  Four 
#>     1     2     3     4
}\if{html}{\out{</div>}}

Functions that start with \code{named} return a subset of the original
object:\if{html}{\out{<div class="r">}}\preformatted{vec <- c(One = 1, Two = 2, Three = 3, Four = 4)
vec \%>\% named("One")
#> One 
#>   1
vec \%>\% named_in(c("Two", "Three"))
#>   Two Three 
#>     2     3
vec \%>\% named_starting("T")
#>   Two Three 
#>     2     3
vec \%>\% named_like("[A-Z].*e$")
#>   One Three 
#>     1     3
}\if{html}{\out{</div>}}

Functions that start with \code{rename} return the object with its names
changed:\if{html}{\out{<div class="r">}}\preformatted{vec \%>\% rename_starting("T", tolower)
#>   One   two three  Four 
#>     1     2     3     4

vec \%>\% rename_in(c("One", "Two"), paste, 1:2, sep = ".")
#> One.1 Two.2 Three  Four 
#>     1     2     3     4

vec \%>\% rename_gsub(2:3, "[aeiou]", "e")
#> Warning in gsub(pattern, replacement, x, ...): argument 'pattern' has length > 1
#> and only the first element will be used
#>       1 [aeiou]       3       4 
#>       1       2       3       4
}\if{html}{\out{</div>}}

Or you can use a one-sided formula, like in
\href{https://purrr.tidyverse.org/}{purrr}:\if{html}{\out{<div class="r">}}\preformatted{vec \%>\% rename_in(c("One", "Two"), ~paste(.x, 1:2, sep = "."))
#> One.1 Two.2 Three  Four 
#>     1     2     3     4
}\if{html}{\out{</div>}}
\subsection{Installation}{

You can install the development version from
\href{https://github.com/}{GitHub} with:\if{html}{\out{<div class="r">}}\preformatted{# install.packages("devtools")
devtools::install_github("hughjonesd/rumpel")
}\if{html}{\out{</div>}}
}
}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com}

}
\keyword{internal}
