% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rumpel-package.R
\docType{package}
\name{rumpel-package}
\alias{rumpel}
\alias{rumpel-package}
\title{Functions to manipulate objects by name}
\description{
In the fairy tale, Rumpelstiltskin could be controlled once the heroine
learned his name. The rumpel package contains convenience functions
to control R objects by name.
}
\details{
Functions that start with \code{named} return a subset of the original
object:\preformatted{vec <- c(one = 1, two = 2, three = 3, four = 4)
vec \%>\% named("one")
vec \%>\% named_like("^t")
}

Functions that start with \code{rename} return the object with its
names changed:\preformatted{vec <- c(one = 1, two = 2, three = 3, four = 4)
vec \%>\% rename_gsub("[aeiou]", "e")
vec \%>\% rename_toupper()
}
}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com}

}
\keyword{internal}
